requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
discord==2.3.2
discord.py==2.5.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
more-itertools==10.6.0
multidict==6.1.0
ollama==0.4.7
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
youtube-transcript-api==1.0.0
